% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fixit.R
\name{glm_fixit}
\alias{glm_fixit}
\title{Misclassification correction of Generalized Linear Model with validation data}
\usage{
glm_fixit(
  formula,
  family = gaussian(),
  data,
  data2,
  proxy_formula = NULL,
  proxy_family = binomial(link = "logit"),
  truth_formula = NULL,
  truth_family = binomial(link = "logit"),
  maxit = 1e+06,
  method = "L-BFGS-B"
)
}
\arguments{
\item{formula}{an object of class "formula". Please refer to \link{formula} for details. This package extends the formula syntax with the "||" symbol to indicate ground truth and proxy variables. For example, "x||w" indicates "w" is a proxy variable of the ground truth variable "x".}

\item{family}{a description of the error distribution and link function to be used in the model. Currently, this function supports \code{\link[=gaussian]{gaussian()}} and \code{\link[=binomial]{binomial()}}.}

\item{data}{a data frame with the primary data}

\item{data2}{a data frame with the validation data}

\item{proxy_formula}{an object of class "formula" to describe the data generating process of the proxy variable. Default to all columns in \code{data2}, i.e. "w ~ ."}

\item{proxy_family}{a description of the error distribution and link function to be used to model the proxy variable. Currently, this function supports \code{\link[=binomial]{binomial()}}.}

\item{truth_formula}{an object of class "formula" to describe the data generating process of the ground truth variable. Default to an intercept only model (we don't know the data generating process), i.e. "x ~ 1"}

\item{truth_family}{a description of the error distribution and link function to be used to model the ground truth variable. Currently, this function supports \code{\link[=binomial]{binomial()}}.}

\item{maxit}{variable get passed to \code{\link[=optim]{optim()}}}

\item{method}{variable get passed to \code{\link[=optim]{optim()}}}
}
\value{
This function returns an object class "glm_fixit"
}
\description{
This function provides the MLE-based misclassification correction method proposed by Carroll.
}
\examples{
## Confusion matrix
table(val_data$w, val_data$x)
## default
glm_fixit(y ~ x || w + z, data = research_data, data2 = val_data)
## a more accurate correction
glm_fixit(y ~ x || w + z, data = research_data, data2 = val_data,
proxy_formula = w ~ x*w*z, truth_formula = x ~ w)
## proxy variable in the outcome
glm_fixit(y || w ~ x + z, data = research_data2, data2 = val_data2)
}
\seealso{
\link{research_data}
}
