% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fixit.R
\docType{data}
\name{research_data}
\alias{research_data}
\alias{val_data}
\title{Simulated data for demonstration}
\format{
An object of class \code{data.frame} with 4700 rows and 3 columns.

An object of class \code{data.frame} with 300 rows and 4 columns.
}
\usage{
research_data

val_data
}
\description{
"research_data" and "val_data" are two datasets for demonstrating this package. Both datasets were generated with the expected regression cofficient for \code{x} being -0.04. The ground truth variable "x" is predicted by the proxy variable "y" with only 60\% accuracy.
}
\keyword{datasets}
