% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{research_data}
\alias{research_data}
\alias{val_data}
\alias{val_data2}
\alias{research_data2}
\title{Synthetic datasets}
\format{
An object of class \code{data.frame} with 4700 rows and 3 columns.

An object of class \code{data.frame} with 300 rows and 4 columns.

An object of class \code{data.frame} with 500 rows and 4 columns.

An object of class \code{data.frame} with 6500 rows and 3 columns.
}
\usage{
research_data

val_data

val_data2

research_data2
}
\description{
\code{research_data}, \code{val_data}, \code{research_data2}, and \code{val_data2} are synthetic datasets for demonstrating how this package works.

\verb{research_data*} datasets represent the primary data wherein a \strong{proxy} variable (\emph{w}) contains misclassifications from an automated classifer. \verb{val_data*} datasets represent the validation data wherein both \strong{groundtruth} (e.g. from human annotation, can either be \emph{x} or \emph{y}) and \emph{w} are included. Theoretically, we are not interested in studying \emph{w} per se. Suppose the groundtruth is \emph{x} and we are interested in the relationship between \emph{x} and \emph{y}; we want to estimate the regression coefficient \eqn{B_{xy}}. But we use \emph{w} as a low-cost proxy of the groundtruth \emph{x} and can only estimate \eqn{B_{wy}}.
}
\details{
The datasets were synthesized with the following parameters:
\subsection{\code{research_data} and \code{val_data}}{

For this pair of datasets, we are interested in the relationship between \code{x} (independent variable) and \code{y} (dependent variable; outcome). The data was synthesized with \eqn{B_{xy}} being 0.
\itemize{
\item \code{w}: proxy variable of \code{x} (independent variable) with misclassifications; the prediction accuracy is 0.6. The prediction error is not differential.
\item \code{y}: dependent variable with no misclassifications
\item \code{z}: another independent variable with no misclassifications
\item \code{x}: groundtruth in the validation data
}
}

\subsection{\code{research_data2} and \code{val_data2}}{
\itemize{
\item \code{w}: proxy variable of \code{y} (dependent variable) with misclassifications; the prediction accuracy is 0.6. The prediction error is differential.
\item \code{x}: independent variable with no misclassifications
\item \code{z}: another independent variable with no misclassifications
\item \code{y}: groundtruth in the validation data
}
}
}
\examples{
\donttest{
## Confusion matrix
table(val_data$w, val_data$x)
## default
glm_fixit(y ~ x || w + z, data = research_data, data2 = val_data)
## proxy variable in the outcome
glm_fixit(y || w ~ x + z, data = research_data2,
data2 = val_data2, family = binomial("logit"))
}
}
\seealso{
\code{\link[=glm_fixit]{glm_fixit()}}
}
\keyword{datasets}
